requirements.txt
snakebite-py3
apache-airflow-providers-celery
apache-airflow-providers-apache-kafka
zope.component
zope.event
zope.hookable
zope.interface
aiohttp
celery
Scrapy
constantly 
incremental 
Automat 
hyperlink 
attrs 
twilio
scrapy_cloudflare_middleware
cfscrape
pymongo[srv]
confluent_kafka 
fastavro
avro 
fastapi[all]
pykafka 
faker
requests 
